[TE-005] refactor: use central cn import
